+++ src/insts/riscv/untils.rs
@@ -10,20 +10,20 @@ pub fn branch(funct: u8, rs1: Reg, rs2: Reg, imm: i16) -> Inst {
 }
 
 
-pub fn load_data(funct: u8, rd: Reg, rs1: Reg, imm: u16) -> Inst {
-    i(0b0000011, rd, funct, rs1, imm)
+pub fn load_data(funct: u8, rd: Reg, rs1: Reg, imm: i16) -> Inst {
+    i(0b0000011, rd, funct, rs1, imm as u16)
 }
 
 #[macro_export]
 macro_rules! ld_impl {
     ($name:ident, $funct:expr) => {
-        pub fn $name(rd: Reg, rs1: Reg, imm: u16) -> Inst {
+        pub fn $name(rd: Reg, rs1: Reg, imm: i16) -> Inst {
             load_data($funct, rd, rs1, imm)
         }
     };
 }
 
-pub fn store_data(funct: u8, rs1: Reg, rs2: Reg, imm: u16) -> Inst {
+pub fn store_data(funct: u8, rs1: Reg, rs2: Reg, imm: i16) -> Inst {
     let imm0_4 = (imm & 0b1111) as u8;
     let imm5_11 = ((imm >> 5) & 0b1111111) as u8;
     s(0b0100011, imm0_4, funct, rs1, rs2, imm5_11)
@@ -32,7 +32,7 @@ pub fn store_data(funct: u8, rs1: Reg, rs2: Reg, imm: u16) -> Inst {
 #[macro_export]
 macro_rules! st_impl {
     ($name:ident, $funct:expr) => {
-        pub fn $name(rs1: Reg, rs2: Reg, imm: u16) -> Inst {
+        pub fn $name(rs1: Reg, rs2: Reg, imm: i16) -> Inst {
             store_data($funct, rs1, rs2, imm)
         }
     };
